Surgery - Add Additional Settings to Allow Patient to Remain Awake During Surgeries (#371)
**When merged this pull request will:**

- Added 1 Dropdown Setting with 4 options In the Surgery category.(In
XEH_preInit)
- Option 1 Surgery Causes Unconsciousness: An incision causes cardiac
arrest if the patient is not unconscious and sedated(With Etomidate).
Uses a Switch Statement in fnc_incisionLocal.sqf
- Checks if the patient Has Etomidate and if they are in an unconscious
state.
- Option 2 Unconsciousness Required for Surgery: Surgery can only begin
if the patient is unconscious and sedated(With Etomidate).
- Changed the First if statement in fnc_incisionLocal from checking for
uncon_requieredForAction to Checking for unconSurgery_requieredForAction
equals 1. This statement checks to make sure the patient is unconscious
and if they aren't then Surgery is aborted(This setting is the old
uncon_requieredForAction, Which has been brought into the switch state
due to the similarity of the job)
- Removed Setting "uncon_requieredForAction" and the localizations for
the same setting
- Option 3 No Unconsciousness from Surgery: The patient is able to stay
awake through the entirety of the surgery, experiencing a little pain
and a slight increase in heart rate (Etomidate not required). Uses a
Switch Statement in fnc_incisionLocal.sqf
- This Case Uses the adjustPainLevel and addMedicationAdjustment, to
cause pain and increase in heart rate[There is a chance this can be
simplified with just the addMedicationAdjustment, but was unsure when
searching for similar use cases in Code so went with 2 separate
functions]
- Option 4 Surgery Anesthesia: The patient can remain awake for the
duration of the surgery. However, if Etomidate hasn't been applied, the
patient will lose consciousness and go into CA. There are two cases in
the switch statement in fnc_incisionLocal.sqf.
- The first case checks the unconSurgery_requieredForAction setting and
whether there's no Etomidate in the patient's system. If so, it induces
cardiac arrest by causing a heart rate increase of 200 and forces the
patient into unconsciousness.
- The second case applies when there's Etomidate in the patient's
system. In this scenario, the incision can proceed without
complications, allowing the surgery to be carried out without issues.
    
- Added English Translations For All new Text Used in Settings Area

---------

Co-authored-by: Cutup <ThatOneGuy63@users.noreply.github.com>
Co-authored-by: Blue <itzblueman@gmail.com>