Surgery - Change Clotting Values (#741)
This PR intends to adjust the clotting duration and reopening chances in
order to reduce the overall effectiveness of clotting and encourage the
use of TXA/EACA for patients with a large number of wounds while still
maintaining a level of effectiveness for patients needing minimal care.

**When merged this pull request will:**
- Lower the minimum duration of natural clots (Minor 300>120, Medium
300>60, Major 150>30)
- Swap the reopening chances of minor and medium natural clots

### IMPORTANT

- [Development Guidelines](https://ace3.acemod.org/wiki/development/)
are read, understood and applied.
- Title of this PR uses our standard template `Component -
Add|Fix|Improve|Change|Make|Remove {changes}`.

---------

Co-authored-by: mazinskihenry <33608576+mazinskihenry@users.noreply.github.com>